// src/app/shared/models/institution.ts
export interface Institution {
    uuid:                string;
    slug:                string;
    name:                string;
    description:         string;
    location:            string;
    category:            string;
    email:               string;
    phone:               string;
    url:                 string;
    logo_url:            string;
    background_url:      string;
    logoFileUuid?:       string;
    backgroundFileUuid?: string;
}